moved CustomEditor in separate file
