+++ packages/components/timed-text-editor/index.js
@@ -2,7 +2,6 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import {
-  Editor,
   EditorState,
   CompositeDecorator,
   convertFromRaw,
@@ -11,8 +10,10 @@ import {
   Modifier
 } from "draft-js";
 
+
+import CustomEditor from './CustomEditor.js';
 import Word from './Word';
-import WrapperBlock from './WrapperBlock';
+
 import sttJsonAdapter from '../../stt-adapters';
 import exportAdapter from '../../export-adapters';
 import updateTimestamps from './UpdateTimestamps/index.js';
@@ -589,61 +590,3 @@ TimedTextEditor.propTypes = {
 };
 
 export default TimedTextEditor;
-
-// TODO: move CustomEditor in separate file
-// NOTE: custom editor is in a separate class to minimise re-renders
-// if considering refactoring, removing the separate class, please double check
-// that doing so does not introduce uncessary re-renders first.
-class CustomEditor extends React.Component {
-  handleWordClick = e => {
-    this.props.onWordClick(e);
-  };
-
-  renderBlockWithTimecodes = () => {
-    return {
-      component: WrapperBlock,
-      editable: true,
-      props: {
-        showSpeakers: this.props.showSpeakers,
-        showTimecodes: this.props.showTimecodes,
-        timecodeOffset: this.props.timecodeOffset,
-        editorState: this.props.editorState,
-        setEditorNewContentState: this.props.setEditorNewContentState,
-        onWordClick: this.handleWordClick,
-        handleAnalyticsEvents: this.props.handleAnalyticsEvents,
-        isEditable: this.props.isEditable
-      }
-    };
-  };
-
-  shouldComponentUpdate(nextProps) {
-    // https://stackoverflow.com/questions/39182657/best-performance-method-to-check-if-contentstate-changed-in-draftjs-or-just-edi
-    if (nextProps.editorState !== this.props.editorState) {
-      return true;
-    }
-
-    if (nextProps.isEditable !== this.props.isEditable) {
-      return true;
-    }
-
-    return false;
-  }
-
-  handleOnChange = e => {
-    this.props.onChange(e);
-  };
-
-  render() {
-    return (
-      <Editor
-        editorState={this.props.editorState}
-        onChange={this.handleOnChange}
-        stripPastedStyles
-        blockRendererFn={this.renderBlockWithTimecodes}
-        handleKeyCommand={this.props.handleKeyCommand}
-        keyBindingFn={this.props.customKeyBindingFn}
-        spellCheck={this.props.spellCheck}
-      />
-    );
-  }
-}
